Update asp-net-troubleshoot-no-data.md
Updated the following to reflect the current version of Application Insights. In the future, it may be a good idea to reword this to reflect a specific version or later version.

The latest version of `Microsoft.ApplicationInsights.AspNetCore` is 2.22.0, and it refers to `Microsoft.ApplicationInsights` version 2.22.0. For this reason, the version of `Microsoft.ApplicationInsights.AspNetCore` to be installed should be 2.22.0.